backend/app/services/ai_services.py
from fastapi import HTTPException, UploadFile
from sqlalchemy.orm import Session
from PIL import Image # For image processing
import io # For handling byte streams
import numpy as np # For numerical operations, image manipulation
from typing import List, Optional # Added Optional for get_fashion_trends_service user

# Placeholder for actual model loading and inference
# from transformers import pipeline # Example: if using Hugging Face transformers
# import tensorflow as tf # Example: if using TensorFlow
# from sklearn.cluster import KMeans # Example: for color quantization

from .. import schemas, models

# --- Mock/Placeholder AI Functions ---
# These will be replaced or augmented with real model calls.

def mock_extract_colors(image: Image.Image, num_colors=5) -> List[str]:
    # Simplified: Resize, get dominant colors (mocked)
    # In a real scenario, use KMeans or similar on pixel data
    # For now, return fixed mock colors based on image size or a random selection
    width, height = image.size
    if width > 500:
        return ["#2563EB", "#DC2626", "#059669", "#F59E0B", "#EF4444"] # Brighter
    else:
        return ["#374151", "#4B5563", "#6B7280", "#9CA3AF", "#D1D5DB"] # Muted
    # Real implementation would involve:
    # image_arr = np.array(image.resize((100, 100))) # Resize for performance
    # pixels = image_arr.reshape(-1, 3)
    # kmeans = KMeans(n_clusters=num_colors, random_state=0, n_init='auto').fit(pixels)
    # dominant_colors = kmeans.cluster_centers_.astype(int)
    # return [f"#{r:02x}{g:02x}{b:02x}" for r, g, b in dominant_colors]

def mock_detect_style(image: Image.Image) -> str:
    # Mock: Based on image aspect ratio or some simple feature
    width, height = image.size
    if width > height:
        return "Casual Landscape"
    elif height > width:
        return "Formal Portrait"
    else:
        return "Modern Square"
    # Real implementation: Use an image classification model (e.g., ResNet, ViT)
    # trained on fashion styles.
    # captioner = pipeline("image-to-text", model="nlpconnect/vit-gpt2-image-captioning")
    # caption = captioner(image)[0]['generated_text']
    # return caption # Or further process caption to extract style

def mock_identify_items(image: Image.Image) -> List[str]:
    # Mock: return fixed items
    # Real implementation: Use an object detection model (e.g., YOLO, Faster R-CNN)
    # trained on clothing items.
    # object_detector = pipeline("object-detection", model="facebook/detr-resnet-50")
    # objects = object_detector(image)
    # return [obj['label'] for obj in objects if obj['score'] > 0.8] # Filter by confidence
    return ["Mock T-Shirt", "Mock Jeans", "Mock Sneakers"]

# --- Main Service Function ---

async def analyze_outfit_image_service(
    file: UploadFile, # Changed from image_data to UploadFile
    db: Session,
    user: schemas.User # Assuming schemas.User from get_current_user
) -> schemas.OutfitAnalysisResponse: # Define this schema
    try:
        image_bytes = await file.read()
        image = Image.open(io.BytesIO(image_bytes)).convert("RGB")
    except Exception as e:
        raise HTTPException(status_code=400, detail=f"Invalid image file: {str(e)}")

    # Simulate AI processing steps
    # In a real application, these would involve calls to ML models/pipelines
    # For Transformers, Scikit-learn, TensorFlow:
    # 1. Color Analysis: Use scikit-learn's KMeans for color quantization from image pixels.
    # 2. Style Detection: Use a pre-trained image classification model from TensorFlow Hub or a Hugging Face Transformer (e.g., ViT - Vision Transformer) fine-tuned on fashion styles.
    # 3. Item Identification: Use an object detection model from TensorFlow Hub (e.g., Faster R-CNN, SSD) or a Hugging Face Transformer (e.g., DETR) fine-tuned on clothing items.
    # 4. Occasion Appropriateness: Could be rule-based initially, or a text classifier on combined item/style info if enough data.

    # Using mock functions for now as per "little time" constraint for initial setup
    # These would be replaced by actual model inference calls
    extracted_colors = mock_extract_colors(image)
    detected_style = mock_detect_style(image)
    identified_items = mock_identify_items(image)

    # Mock recommendations and insights
    recommendations = [
        "Consider adding a statement accessory.",
        "This style is great for casual Fridays.",
    ]
    if "Mock T-Shirt" in identified_items:
        recommendations.append("A patterned t-shirt could be more expressive.")

    # For demonstration, we're not saving the analysis to DB yet, but one might
    # create an OutfitAnalysis model and store results linked to the user.

    return schemas.OutfitAnalysisResponse(
        fileName=file.filename,
        contentType=file.content_type,
        style=detected_style,
        dominantColors=extracted_colors, # Renamed from colors
        identifiedItems=identified_items,
        occasionSuitability="General Casual", # Mocked
        confidenceScore=0.85, # Mocked
        recommendations=recommendations,
        # colorPalette and styleInsights from frontend mock can be added if needed
        colorPalette=[{"color": c, "name": "Color " + c, "percentage": round(100/len(extracted_colors),1)} for c in extracted_colors],
        styleInsights=[{"category": "Overall", "score": 85, "description": f"The outfit leans towards {detected_style}."}]

    )

# Placeholder for trend forecasting service (will be in Step 4)
# In backend/app/services/ai_services.py
# Replace the existing placeholder for get_fashion_trends_service

# ... (other imports and functions like analyze_outfit_image_service) ...
from typing import Optional # Make sure Optional is imported from typing

async def get_fashion_trends_service(db: Session, user: Optional[schemas.User] = None) -> schemas.TrendForecastResponse:
    # Mock implementation for now, structured like AITrendForecasting.tsx's mockForecast
    # In a real scenario:
    # - Fetch raw trend data (e.g., from social media, fashion articles, sales data).
    # - Process with NLP (transformers) for text data.
    # - Use time-series analysis (sklearn, tensorflow) for sales/popularity data.
    # - Generate personalized recommendations based on user's wardrobe/style history if 'user' is provided.

    mock_trends_data = [
        schemas.TrendDataItem(
            id='1',
            name='Neo-Cottagecore AI',
            category='Aesthetic Movement (Backend)',
            popularity=94,
            growth=267,
            description='Backend AI: A futuristic take on cottagecore with sustainable tech fabrics and earthy tones.',
            colors=['Sage Green', 'Mushroom Brown', 'Lavender', 'Cream'],
            season='Spring 2025',
            confidence=96,
            outfitImages=['https://via.placeholder.com/150/A2D2A2/000000?text=NeoCottage1', 'https://via.placeholder.com/150/D2A2D2/000000?text=NeoCottage2'],
            celebrities=['Celeb A (AI)', 'Celeb B (AI)'],
            hashtags=['#NeoCottagecoreAI', '#SustainableFashionAI'],
            priceRange='$100-$350',
            occasion=['Casual AI', 'Work From Home AI']
        ),
        schemas.TrendDataItem(
            id='2',
            name='Cyber-Minimalism AI',
            category='Fashion Tech (Backend)',
            popularity=89,
            growth=189,
            description='Backend AI: Clean lines meet smart fabrics with embedded tech and holographic accents.',
            colors=['Chrome Silver', 'Deep Black', 'Electric Blue', 'Pure White'],
            season='Fall 2025',
            confidence=92,
            outfitImages=['https://via.placeholder.com/150/C0C0C0/FFFFFF?text=CyberMin1', 'https://via.placeholder.com/150/0000FF/FFFFFF?text=CyberMin2'],
            celebrities=['Celeb C (AI)', 'Celeb D (AI)'],
            hashtags=['#CyberMinimalAI', '#TechWearAI'],
            priceRange='$180-$850',
            occasion=['Night Out AI', 'Creative Events AI']
        )
    ]

    mock_personalized_recommendations = schemas.PersonalizedRecommendations(
        mustHave=['AI: Smart casual blazer', 'AI: Sustainable denim'],
        avoid=['AI: Fast fashion basics', 'AI: Overly specific trends'],
        investIn=['AI: Quality tech outerwear', 'AI: Versatile pieces']
    )

    mock_seasonal_predictions = schemas.SeasonalPredictions(
        emerging=['AI: Biometric jewelry', 'AI: Climate-responsive clothing'],
        declining=['AI: Logo-heavy designs', 'AI: Single-use items'],
        stable=['AI: Classic denim', 'AI: White shirts (backend)']
    )

    return schemas.TrendForecastResponse(
        trends=mock_trends_data,
        personalizedRecommendations=mock_personalized_recommendations,
        seasonalPredictions=mock_seasonal_predictions
    )
