backend/app/tests/test_ai_services.py
import pytest
from fastapi import UploadFile
from sqlalchemy.orm import Session # For type hinting if needed for mock db
from PIL import Image
import io
import os # For creating dummy files if necessary

from ...app.services import ai_services # Relative import from tests directory
from ...app import schemas # Relative import for schemas

# Mock a database session if needed by services (though our current mock AI services don't use it heavily)
class MockDBSession:
    def query(self, *args, **kwargs): return self # Simplistic mock
    def filter(self, *args, **kwargs): return self
    def first(self): return None
    def all(self): return []
    def add(self, *args, **kwargs): pass
    def commit(self): pass
    def refresh(self, *args, **kwargs): pass

# Mock current_user schema object
# For Pydantic v2, ensure datetime strings are valid or use datetime objects.
# Using valid ISO 8601 strings for datetimes as an example.
mock_user_data = schemas.User(
    id=1,
    username="testuser",
    email="test@example.com",
    created_at="2023-01-01T12:00:00Z",
    updated_at="2023-01-01T12:00:00Z"
)


@pytest.mark.asyncio
async def test_analyze_outfit_image_service_mock_logic():
    # Create a dummy image file in memory
    img_byte_arr = io.BytesIO()
    image = Image.new('RGB', (600, 400), color = 'red') # Create a dummy image
    image.save(img_byte_arr, format='PNG')
    img_byte_arr.seek(0) # Move to the start of the BytesIO object

    # Create a mock UploadFile
    mock_file = UploadFile(filename="test_image.png", file=img_byte_arr, content_type="image/png")

    mock_db_session = MockDBSession()

    analysis_result = await ai_services.analyze_outfit_image_service(
        file=mock_file,
        db=mock_db_session,  # Pass the mock session
        user=mock_user_data
    )

    assert analysis_result is not None
    assert analysis_result.fileName == "test_image.png"
    assert analysis_result.contentType == "image/png"
    assert "style" in analysis_result.model_fields_set # Pydantic v2: model_fields_set
    assert "dominantColors" in analysis_result.model_fields_set
    assert "identifiedItems" in analysis_result.model_fields_set

    # Check mock logic outputs (these might change if mock logic is updated)
    # Example: based on current mock_extract_colors for an image > 500px width
    assert "#2563EB" in analysis_result.dominantColors
    # Example: based on current mock_detect_style for width > height
    assert analysis_result.style == "Casual Landscape"
    assert "Mock T-Shirt" in analysis_result.identifiedItems

    print(f"Analysis Result: {analysis_result.model_dump_json(indent=2)}")


@pytest.mark.asyncio
async def test_get_fashion_trends_service_mock_logic():
    mock_db_session = MockDBSession()
    trends_result = await ai_services.get_fashion_trends_service(db=mock_db_session, user=mock_user_data)

    assert trends_result is not None
    assert len(trends_result.trends) > 0
    assert trends_result.trends[0].name == 'Neo-Cottagecore AI' # Check one of the mock trends
    assert "mustHave" in trends_result.personalizedRecommendations.model_fields_set
    assert "emerging" in trends_result.seasonalPredictions.model_fields_set

    print(f"Trends Result: {trends_result.model_dump_json(indent=2)}")

# To run these tests (from the `backend` directory):
# Ensure pytest and pytest-asyncio are installed: pip install pytest pytest-asyncio
# Command: pytest
# (You might need to set PYTHONPATH=. or python -m pytest)
